Warn on unsupported test macros in ParseAndAddCatchTests
Handling all the possible test macros is nearly infeasable for a simple helper
script written in CMake. But we don't want to silently miss test cases, while
the user thinks that all is thoroughly tested.

To prevent that warn the user when such known unhandled test macros are found
while parsing the source file.

Introduce a new option to enable/silence the unsupported macro warnings
`PARSE_CATCH_TESTS_UNSUPPORTED_WARNINGS` defaulting to 'ON'.